Combine test suites
